Initial implementation of generic error handling
